+++ packages/icons/luminous_symbols_icons/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+- Rolled the Luminous Symbols font from `v42` to `v46`.
+  - This update introduces 33 new icons: `apps`, `auto_delete`, `bookmark`, `bookmarks`, `brightness_7`, `build`, `collections`, `draft`, `draft_spark`, `feedback`, `flight`, `flowchart`, `g_translate`, `google_app`, `history_off`, `home_2`, `incognito`, `notes_spark`, `notifications`, `person_off`, `pets`, `preview`, `search_off`, `sentiment_dissatisfied`, `swipe_down`, `terminal`, `text_fields`, `translate`, `videocam_off`, `visibility`, `warning`, `wb_twilight`, `widget_menu`.
+
 ## 0.1.1
 
 - Rolled the Luminous Symbols font from `v38` to `v42`.